Move analytics effects to separate folder
