+++ src/navigation/card/components/CardDashboard.tsx
@@ -28,8 +28,8 @@ import {CardProvider} from '../../../constants/card';
 import {CARD_WIDTH} from '../../../constants/config.card';
 import {navigationRef} from '../../../Root';
 import {AppEffects} from '../../../store/app';
+import {Analytics} from '../../../store/analytics/analytics.effects';
 import {showBottomNotificationModal} from '../../../store/app/app.actions';
-import {Analytics} from '../../../store/app/app.effects';
 import {selectBrazeCardOffers} from '../../../store/app/app.selectors';
 import {CardEffects} from '../../../store/card';
 import {Card, UiTransaction} from '../../../store/card/card.models';
