+++ src/store/bitpay-id/bitpay-id.effects.ts
@@ -12,7 +12,8 @@ import Dosh from '../../lib/dosh';
 import {isAxiosError, isRateLimitError} from '../../utils/axios';
 import {generateSalt, hashPassword} from '../../utils/password';
 import {AppActions, AppEffects} from '../app/';
-import {Analytics, startOnGoingProcessModal} from '../app/app.effects';
+import {Analytics} from '../analytics/analytics.effects';
+import {startOnGoingProcessModal} from '../app/app.effects';
 import {CardEffects} from '../card';
 import {Effect} from '../index';
 import {LogActions} from '../log';
