+++ src/store/card/card.effects.ts
@@ -7,7 +7,8 @@ import {batch} from 'react-redux';
 import CardApi from '../../api/card';
 import {InitialUserData} from '../../api/user/user.types';
 import {sleep} from '../../utils/helper-methods';
-import {Analytics, startOnGoingProcessModal} from '../app/app.effects';
+import {Analytics} from '../analytics/analytics.effects';
+import {startOnGoingProcessModal} from '../app/app.effects';
 import {Effect} from '../index';
 import {LogActions} from '../log';
 import {ProviderConfig} from '../../constants/config.card';
